+++ startup/GafferCyclesUI/cyclesExamples.py
@@ -23,6 +23,16 @@ GafferUI.Examples.registerExample(
 )
 
 GafferUI.Examples.registerExample(
+	"Rendering/Anamorphic Camera",
+	"$GAFFER_ROOT/resources/examples/rendering/anamorphicCameraSetup.gfr",
+	description = "Demonstrates how to set up a Camera with a non-square pixel aspect ratio.",
+	notableNodes = [
+		GafferScene.Camera,
+		GafferScene.StandardOptions
+	]
+)
+
+GafferUI.Examples.registerExample(
 	"Rendering/Macbeth Chart",
 	"$GAFFER_ROOT/resources/examples/rendering/macbethChart.gfr",
 	description = "Demonstrates how to create and assign a procedural OSL texture.",