+++ Changes.md
@@ -37,6 +37,7 @@ Build
 Documentation
 -------------
 
+- Updated the "Anamorphic Camera Setup" example to use Cycles instead of Appleseed.
 - Updated the "Contact Sheet Generation" example to use Cycles instead of Appleseed.
 - Updated the "Macbeth Chart" example to use Cycles instead of Appleseed.
 - Updated the "Per-location Light Tweak Spreadsheet" example to use Cycles instead of Appleseed.